requirements.txt
python==3.9.0
django==4.0.3
djangorestframework==3.13.1